model patcher promoted to own class
